api/requirements.txt
djangorestframework==3.10.3
Django==3.0
djangorestframework-jwt
drf-nested-routers==0.90.2